refactor: main 패키지 구조 변경에 맞춰 test 패키지 구조 동기화
