+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -1,7 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import lotto.domain.LottoResult;
-import lotto.domain.Rank;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
